--- settings.gradle
@@ -1,4 +1,0 @@
-if (new File("oblivium").exists()) {
-    includeBuild("oblivium/sgx-jvm-plugin")
-    includeBuild("oblivium")
-}
