+++ src/main/kotlin/com/r3/sgx/enclave/rng/RngEnclave.kt
@@ -2,15 +2,16 @@ package com.r3.sgx.enclave.rng
 
 import com.r3.sgx.core.common.*
 import com.r3.sgx.core.enclave.EnclaveApi
-import com.r3.sgx.core.enclave.EnclaveletEnclave
-import com.r3.sgx.core.common.SchemesSettings
-import com.r3.sgx.core.common.SignatureScheme
+import com.r3.sgx.core.enclave.Enclavelet
 import java.nio.ByteBuffer
 import java.security.KeyPair
 import java.security.MessageDigest
-import kotlin.math.min
 
-class RngEnclave : EnclaveletEnclave() {
+/**
+ * A Random Number Generator [Enclavelet]. It creates an EDDSA key-pair and puts the hash of the public key into the
+ * enclave's report. Clients are then handled by [RngHandler]
+ */
+class RngEnclave : Enclavelet() {
     private lateinit var signatureScheme: SignatureScheme
     private lateinit var keyPair: KeyPair
 
@@ -26,27 +27,3 @@ class RngEnclave : EnclaveletEnclave() {
     }
 }
 
-class RngHandler(val keyPair: KeyPair, val signatureScheme: SignatureScheme, val api: EnclaveApi): Handler<Sender> {
-    override fun receive(connected: Sender, input: ByteBuffer) {
-        val sizeRequested = input.getInt()
-        val randomBytesSize = min(sizeRequested, 1024)
-        val randomBytes = ByteArray(randomBytesSize)
-        api.getRandomBytes(randomBytes, 0, randomBytesSize)
-        val signature = signatureScheme.sign(keyPair.private, randomBytes)
-        val publicKey = keyPair.public.encoded
-        // We could use something like protobuf, but for the sake of reducing build dependencies we do custom serialization
-        val size = 4 + randomBytesSize + 4 + publicKey.size + 4 + signature.size
-        connected.sendTopLevel(size) {
-            it.putInt(randomBytesSize)
-            it.put(randomBytes)
-            it.putInt(publicKey.size)
-            it.put(publicKey)
-            it.putInt(signature.size)
-            it.put(signature)
-        }
-    }
-
-    override fun connect(upstream: Sender): Sender {
-        return upstream
-    }
-}
